Refactor migrate-libs
